Refactor engine and make it a session manager
This update makes engine a session manager instead of a single-game
handler, as well as improves the data supplied by `GameUpdate` and
refactors engine to accomodate those changes.

So far, engine was a representation for a single game session, such that
each instance was resposible for a specifc session, and a new instance
had to be created for a new session.
This approach had 2 problems:
- It would have made UI implementations feel a bit clunky, as they
  usually allow for playing multiple sessions.
- It forced `Engine`'s interface to be somewhat leaky: it was mainly
  event-driven, but the state of the game at the start of every session
  could not be delivered in such way - because no event has happened
  directly after listeners were attached and before making a move.
Turning engine into a session manager solves both those problem.

Other notable changes include:
- Made `GameUpdate` fuller by adding fields for session and offered
  draw, replacing `query` with full `GameState`,
  and making `in_check` a derived method instead of a field.
- Made all player action methods non-silent: they now return an
  appropriate error whenever the action wasn't processed,
  e.g. due to offering a draw twice.
- Add a private method `Engine#update_game` as the single entry point
  for all game and session updates.
- replaced public `#query` with `#last_udate`, which includes the full
  latest successful `GameUpdate`.